# docs/swagger.yaml
basePath: /api/v2
definitions:
  apimodels.Comment:
    properties:
      authorId:
        type: integer
      id:
        type: integer
      message:
        type: string
      notifyId:
        type: integer
    type: object
...
  apimodels.MessageResponse:
    properties:
      message:
        type: string
    type: object
...
  apimodels.PostComment:
    properties:
      message:
        type: string
      notifyId:
        type: integer
    type: object
...
  apimodels.SuccessPostComment:
    properties:
      id:
        type: integer
    type: object
...
info:
  contact: {}
  description: Swagger API for Dreamit-based web labs
  termsOfService: http://swagger.io/terms/
  title: Dreamit Swagger API
  version: "2.0"
paths:
  /comments:
    get:
      description: Get notify comments collection
      parameters:
      - description: Notify to filter comments for, obligatory
        in: query
        name: notifyId
        required: true
        type: integer
      - description: token with the bearer started
        in: header
        name: Authorization
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/apimodels.Comment'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apimodels.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apimodels.MessageResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apimodels.MessageResponse'
      security:
      - ApiKeyAuth: []
      summary: Get notify comments collection
      tags:
      - comment
    post:
      consumes:
      - application/json
      description: Create comment and get id
      parameters:
      - description: New comment
        in: body
        name: id
        required: true
        schema:
          $ref: '#/definitions/apimodels.PostComment'
      - description: token with the bearer started
        in: header
        name: Authorization
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/apimodels.SuccessPostComment'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apimodels.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apimodels.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apimodels.MessageResponse'
      security:
      - ApiKeyAuth: []
      summary: Create new comment
      tags:
      - comment
  /comments/{id}:
    delete:
      description: Delete comment by id
      parameters:
      - description: Comment ID
        in: path
        name: id
        required: true
        type: integer
      - description: token with the bearer started
        in: header
        name: Authorization
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apimodels.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apimodels.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apimodels.MessageResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete comment
      tags:
      - comment
    get:
      description: Get comment by id
      parameters:
      - description: Comment ID
        in: path
        name: id
        required: true
        type: integer
      - description: token with the bearer started
        in: header
        name: Authorization
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apimodels.Comment'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apimodels.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apimodels.MessageResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apimodels.MessageResponse'
      security:
      - ApiKeyAuth: []
      summary: Get comment
      tags:
      - comment
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"